+++ src/utils/html.ts
@@ -1,0 +1,3 @@
+export function containsHtmlMarkup(value: unknown): value is string {
+  return typeof value === 'string' && /<\/?[a-z][^>]*>/i.test(value);
+}